Architecture overhaul

+++ SimpleTerminal/PrintTable.cpp
@@ -9,7 +9,7 @@ PrintTable::PrintTable(const std::vector<std::string>& vHeaders) :
 	m_vColumnWidths.resize(m_vHeaders.size());
 	for (int i = 0; i < m_vHeaders.size(); ++i)
 	{
-		m_vColumnWidths[i] = m_vHeaders[i].size();
+		m_vColumnWidths[i] = static_cast<int>(m_vHeaders[i].size());
 	}
 }
